progress on commands & auto-complete and syntax highlighting now working
