+++ pses_basis/src/command.cpp
@@ -1,0 +1,5 @@
+#include <pses_basis/command.h>
+
+Command::Command(const CommandParams& cmdParams) {
+
+}
