+++ pses_basis/src/serialinterface.cpp
@@ -1,4 +1,4 @@
-#include "pses_basis/serialinterface.h"
+#include <pses_basis/serialinterface.h>
 
 SerialInterface::SerialInterface()
 {
